Simplify SolidIndexes implementation
The "SolidIndexes" method is used to deal with "solid" archives, which are
archives that are compressed as a whole (so accessing an individual file
requires first decompressing every single file prior to it.)
It searches for largest value in the array, and creates a new array of all
indices from 0 to N, inclusive.

This method is always handed a pre-sorted array, so to get the largest value,
all that's needed is to return the last element.